requirement.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-blob==12.17.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
datasets==2.16.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.6
dm-tree==0.1.8
dominate==2.8.0
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.2
fastapi-responses==0.2.1
filelock==3.12.2
firebase-admin==6.2.0
Flask==2.3.2
flask-babel==3.1.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.4
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
gast==0.5.4
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.20.1
idna==3.4
imbalanced-learn==0.11.0
ipykernel==6.25.0
ipython==8.14.0
iso8601==2.1.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jwcrypto==1.5.0
keras==2.14.0
keras-nightly==3.0.4.dev2024020503
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mysqlclient==2.2.0
namex==0.0.7
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
Pyrebase4==4.7.1
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.8.2
python-jwt==4.0.0
python-socketio==5.11.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.6.0
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rich==13.7.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
setfit==1.0.3
simple-websocket==1.0.0
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
SQLAlchemy==2.0.16
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tb-nightly==2.16.0a20240205
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-nightly==2.16.0.dev20240202
tf_keras-nightly==2.16.0.dev2024020510
tf_nightly_intel==2.16.0.dev20240202
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.36.2
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
virtualenv==20.23.1
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
